refactor: move state out of translation context
